+++ packages/plots/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.1.3
+
+`2023-12-12`
+
+- 💄 代码逻辑优化
+- 🐞 修复 DualAxes Annotations 配置展示异常
+
+
 ## 2.1.2
 
 `2023-12-12`
