+++ packages/plots/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ant-design/plots",
-  "version": "2.1.2",
+  "version": "2.1.3",
   "description": "G2Plot Statistical chart",
   "bugs": {
     "url": "https://github.com/ant-design/ant-design-charts/issues"
