+++ packages/util/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ant-design/charts-util",
-  "version": "0.0.1-alpha.4",
+  "version": "0.0.1-alpha.5",
   "description": "charts utils",
   "bugs": {
     "url": "https://github.com/ant-design/ant-design-charts/issues"
